Added popular category box to event page
Added the popular box to the event page.  This gets the popular events from the
category that the current event is in.  The code to get categories had to be
modified to accept a given category instead of just the current category the
user is visiting because all of the CATEGORY constants are set to blank when
viewing an event.  This is because init.php gets the category from the URL
which doesn't make sense on an event page.  Unfortunately the CATEGORY
constants will be set to blank during this process because some code/ needs
them.

The sidecol was also added to the event page so that we can fit multiple items
on the right side, on top of each other.

Fixed some HTML formatting inconsistencies.

Basically just fixed the indenting and added a few comments to mark the end of
divs.