BUG[15]: create captcha
All forms now have 2 additional text fields on them with the class of nospam.
The first thing that ap_basic.js does is give all objects with the nospam class
the class of nobots.  The nobots class has a display: hidden which will keep
people.  One of the text fields must remain blank and the other must keep the ap
value.