BUG[42]: menu in IE loads fucked up
IE8 HATES dd_roundies.  dd_roundies seems to remove any size properties from
spans causing the links to visibly move to the top left once the page is loaded.
I've replaced dd_roundies with jquery.corner which seems to be much better and
works in all browsers (still need to test older versions of ie).  The sub-menu
categories also seemed to load behind the date menu so I gave the date menu a
-1 z-index.  I read that giving the menu li's a higher index than the submenu
ul's/li's should fix it but that didn't seem to work for me (the examples i saw
were divs not lists).

seemed to load in