BUG[42]: menu in IE loads fucked up
After changing the z-index of the date menu to -1 the links were no longer
clickable.  Found that the category menu had a div with a relative position that
wrapped the ul which has a relative position.  I had to give the div a higher
z-index to get the fix mentioned in the last commit to work.  Menus now work
great.

	I read that giving the menu li's a higher index than
	the submenu ul's/li's should fix it but that didn't
	seem to work for me (the examples i saw were divs not
	lists).