BUG[14]: Need a footer with links to pages like help, about, site map, advertising, joinpages
Added a footer to the bottom of the page.  Remove the body line-height to make
it flush with the bottom of the window.

Added an about page.  It's just a short 2 paragraphs and needs some work but its
good enough for now

Added a help page with content.  Could use some more stuff such as:
How to share events with facebook/twitter?
What are these popular events?
Comments.

Added a link for feedback and a blank page.  Need to turn this into a modal form